Refactor: Extract to NPlusOneQueryMethods
