+++ lib/rubocop/cop/isucon_cops.rb
@@ -5,6 +5,7 @@ require_relative "isucon/mixin/join_without_index_methods"
 require_relative "isucon/mixin/offense_location_methods"
 require_relative "isucon/mixin/many_join_table_methods"
 require_relative "isucon/mixin/mysql2_xquery_methods"
+require_relative "isucon/mixin/n_plus_one_query_methods"
 require_relative "isucon/mixin/select_asterisk_methods"
 require_relative "isucon/mixin/sinatra_methods"
 require_relative "isucon/mixin/sqlite3_execute_methods"
